extract functions into individual files, reduce nesting in WordInfo function
